Task user_session:close_expired: fix EXECUTE honoring

+++ lib/vpsadmin/api/tasks/user_session.rb
@@ -23,7 +23,7 @@ module VpsAdmin::API::Tasks
             )
           end
           
-          t.destroy!
+          t.destroy! if ENV['EXECUTE'] == 'yes'
         end
       end
     end
